update script color

+++ tc_coverity.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 export CODE_BASE_DIR=`pwd`
+
 HERE=$(dirname $(realpath $0))
 IDIR_DIR=$CODE_BASE_DIR/idir
 TC_COVERITY_DIR=$HERE/tc_coverity
@@ -8,7 +9,12 @@ COVERITY_PLUGIN_DIR=$HOME/.synopsys/desktop/controller/logs
 CONFIGS_DIR=$TC_COVERITY_DIR/configs/latest-release
 COVERITY_ID_PASS="telechips07"
 
-echo CONFIGS_DIR : $CONFIGS_DIR
+source $HERE/colors.sh
+
+echo -e ${LGreen}[INFO] Base Code Directory${NC} : ${Blue}$CODE_BASE_DIR${NC}
+echo -e ${LGreen}[INFO] Config Directory${NC} : ${Blue}$CONFIGS_DIR${NC}
+
+
 
 STREAM_ID=
 
@@ -19,21 +25,17 @@ if [ -d $COVERITY_PLUGIN_DIR ]; then
 
     PLUGIN_IDIR_PATH=${PLUGIN_IDIR_PATH/idir_full/idir}
 
-    echo PLUGIN_IDIR_PATH : $PLUGIN_IDIR_PATH
+    echo -e ${LGreen}PLUGIN_IDIR_PATH${NC} : ${Blue}$PLUGIN_IDIR_PATH${NC}
 else 
-    echo "No COVERITY_PLUGIN_DIR"
+    echo -e "${Yellow}[WARRNING] No coverity configs dir for vscode${NC}"
 fi
 
-echo Base Code Directory : $CODE_BASE_DIR
-
-RED='\e[0;31m'
-NC='\e[0m' # No Color
-GREEN='\e[0;33m'
 
 if [ -f $CODE_BASE_DIR/tmp.conf ]; then
     source $CODE_BASE_DIR/tmp.conf
-    echo BUILD_CMD : $BUILD_CMD
-    echo STREAM_ID : $STREAM_ID
+    echo -e ${LGreen}Module${NC} : ${Blue}$MODULE_NAME${NC}
+    echo -e ${LGreen}BUILD_CMD${NC} : ${Blue}$BUILD_CMD${NC}
+    echo -e ${LGreen}STREAM_ID${NC} : ${Blue}$STREAM_ID${NC}
 fi
 
 # get options:
@@ -60,13 +62,13 @@ while (( "$#" )); do
 
                 if [ $CONFIG_INDEX -eq $INDEX ]; then exit 1; fi
 
-                echo -e "Selected config file is ${GREEN}"${array[$CONFIG_INDEX]}"${NC}"
+                echo -e "Selected config file is ${Green}"${array[$CONFIG_INDEX]}"${NC}"
 
                 cp -r $TC_COVERITY_DIR/${array[$CONFIG_INDEX]} $CODE_BASE_DIR/tmp.conf
                 cp -r $TC_COVERITY_DIR/coverity.conf $CODE_BASE_DIR/coverity.conf
 
                 if [ -d $CODE_BASE_DIR/tc_coverity ]; then
-                    echo "already has tc_coverity dir, re-set up"
+                    echo -e "${Yellow}[WARNNING]already has tc_coverity dir, re-set up${NC}"
                     rm -rf $CODE_BASE_DIR/tc_coverity
                 fi
                 
@@ -79,44 +81,58 @@ while (( "$#" )); do
                 fi
 
             else   
-                echo "${RED}[Error] there are no conf file in tc_coverity directory${NC}"
+                echo -e "${Red}[Error] there are no conf file in tc_coverity directory${NC}"
             fi
             exit 1
             ;;
         -s|--setup)
-            if [ -f $CODE_BASE_DIR/coverity.conf ]; then
-                if [ -d $IDIR_DIR ]; then
-                    rm -rf $IDIR_DIR
-                fi
+            #if [ -f $CODE_BASE_DIR/coverity.conf ]; then
+            if [ -d $IDIR_DIR ]; then
+                rm -rf $IDIR_DIR
+            fi
 
-                $HERE/clean_lmf.sh $MODULE_NAME
-                
-                BUILD="cov-build --dir $IDIR_DIR  --emit-complementary-info --config $TC_COVERITY_DIR/lmf_coverity_config/coverity_configure_lmf.xml $BUILD_CMD"
-                
-                eval $BUILD
-                
+            $HERE/clean_lmf.sh $MODULE_NAME
+            
+            BUILD="cov-build --dir $IDIR_DIR  --emit-complementary-info --config $TC_COVERITY_DIR/lmf_coverity_config/coverity_configure_lmf.xml $BUILD_CMD"
+            
+            echo -e "${Green}[INFO] #########BUILD START###########${NC}"
+
+            eval $BUILD
+
+            echo -e "${Green}[INFO] #########BUILD FINISH###########${NC}"
+            
+            if [ -d $COVERITY_PLUGIN_DIR ]; then
                 if [ -d $PLUGIN_IDIR_PATH ]; then
                     rm -rf $PLUGIN_IDIR_PATH
                 fi
-                echo "Linking in plugins dir $IDIR_DIR"
-                echo "to $PLUGIN_IDIR_PATH"
+                echo -e "${Green}[INFO] Try to link at plugins dir $IDIR_DIR${NC}"
+                echo -e "${Green}to${NC} ${Blue}$PLUGIN_IDIR_PATH${NC}"
                 ln -s $IDIR_DIR $PLUGIN_IDIR_PATH
-                exit 1
-            else
-                echo "Error: the coveirty.conf file isn't exist"
-                exit 1
             fi
+
+            exit 1
+            #else
+            #    echo -e "${Red}[ERROR]Error: the coveirty.conf file isn't exist${NC}"
+            #    exit 1
+            #fi
             ;;
         -l|--link)
+
+            if [ -d $COVERITY_PLUGIN_DIR ]; then
+                echo -e "${Red}[ERROR] No coverity configs dir for vscode${NC}"
+                exit 1
+            fi
+
             if [ -d $PLUGIN_IDIR_PATH ]; then
                 rm -rf $PLUGIN_IDIR_PATH
             fi
+
             if [ ! -d $IDIR_DIR ]; then
-                echo "Not exist builded Idir dir" 
+                echo -e "${Red}Not exist builded Idir dir${NC}" 
                 exit 1
             fi
-            echo "Linking in plugins dir $IDIR_DIR"
-            echo "to $PLUGIN_IDIR_PATH"
+            echo -e "${Green}[INFO] Try to link at plugins dir $IDIR_DIR${NC}"
+            echo -e "${Green}to${NC} ${Blue}$PLUGIN_IDIR_PATH${NC}"
             ln -s $IDIR_DIR $PLUGIN_IDIR_PATH
             exit 1
             ;;
@@ -129,10 +145,12 @@ while (( "$#" )); do
                 --coding-standard-config $CONFIGS_DIR/cert-c-recommendation-telechips-210714.config \
                 --config $TC_COVERITY_DIR/lmf_coverity_config/coverity_configure_lmf.xml \
                 @@$CONFIGS_DIR/runtime_rules_telechips_211119.txt"
-                    COV_ANALYZE_OPTIONS="$COV_ANALYZE_OPTIONS --parse-warnings-config $CONFIGS_DIR/parse_warnings_telechips_211119.conf"
+                
+                COV_ANALYZE_OPTIONS="$COV_ANALYZE_OPTIONS --parse-warnings-config $CONFIGS_DIR/parse_warnings_telechips_211119.conf"
+                
                 cov-analyze $COV_ANALYZE_OPTIONS
             else
-                echo "the captured directory not exist"
+                echo -e "${Red}[ERROR] The captured directory does not exist${NC}"
             fi
             exit 1
             ;;
@@ -144,14 +162,14 @@ while (( "$#" )); do
                     cov-commit-defects --dir $IDIR_DIR --url http://coverity.telechips.com:8080 --user $COVERITY_ID_PASS --password $COVERITY_ID_PASS --stream $STREAM_ID
                 fi
             else
-                echo "the captured directory not exist"
+                echo -e "${Red}[ERROR] The captured directory does not exist${NC}"
             fi
             exit 1
             ;;
         -f|--filter)
             if  [ $3 == "-s" ] || [ $3 == "--stream" ]; then
                 STREAM_ID=$4
-                echo "set stream to $STREAM_ID"
+                echo -e "set stream to ${Green}$STREAM_ID${NC}"
             fi
             if [ $2 == "get" ]; then 
                 cov-manage-findings --dir $IDIR_DIR --stream $STREAM_ID --url http://coverity.telechips.com:8080 --user $COVERITY_ID_PASS --password $COVERITY_ID_PASS --action readFromConnect --report my_findings_report_output.xlsx
@@ -167,9 +185,10 @@ while (( "$#" )); do
             exit 1
             ;;
          -h|--help)
+             echo "[option] tmp.conf : configs (Moduel name, make cammand, stream id"
              echo "[option] config -> setup -> analysis -> commit"
-             echo " -c : config (select : cdk-audio, libomxil-telechips, gst-plugins-telechips)"
-             echo " -s : setup (cleanall, build)"
+             echo " -c : config (select : dvrs, vpu_k54, libcdk-audio, libomxil-telechips, gstreamer1.0-plugins-telechips)"
+             echo " -s : setup (clean and build)"
              echo " -a : analysis (config dir: $CONFIGS_DIR)"
              echo " -e : commit (parsing from converity.conf, if it has addtional stream_id then will be applied"
              echo " -f : filter"
@@ -184,7 +203,7 @@ while (( "$#" )); do
             exit 1
             ;;
          *)
-            echo "Not support command"
+            echo -e "${Red}[ERROR] Not support command${NC}"
             exit 0
             ;;
     esac
